Add a message queue to Writer
I believe this will fix client crashing when big files are being sent. A big file would get sent and while that's being sent I write a ping to the client and it writes it to the file.

Now pings don't get sent at all and messages are queued while a file is being sent.